set dynamic block access modal
